修复closetime值不对问题
时间初始化不对，开始日期时间不能为0，Year里面存的值超过255，改为uint16_t 类型

+++ User/screen/screenStatus.h
@@ -11,7 +11,7 @@
 
 typedef struct
 {
-	uint8_t Year;
+	uint16_t Year;
 	uint8_t Mon;
 	uint8_t Day;
 	uint8_t Hour;
